VolumePanel: Properly update expanded volume panel sliders
Fixed a series of bugs where volume sliders are not updated when
toggling ringer modes or changing Zen mode with the expanded volume
panel

updateStates() on line 720 is needed to enable (resp. disable) the
sliders when volume is unmuted (resp. muted) by clicking the icon.

On lines 1593--1597, we need to check whether the volume panel is
expanded: if so, changing Zen modes may require updating more than
just the active stream's slider.

updateStates() on line 1682 is needed to enable the sliders when
volume is unmuted by dragging the slider from muted state.

Change-Id: I3f6c01320a53b6ea37abae9145d7d520d3ea46e0

Conflicts:

	packages/SystemUI/src/com/android/systemui/volume/VolumePanel.java