Settings: Sounds/Ringer Icon consistency change
Just like in Settings for audio_icon.xml, when you enable and disable
"Increasing ring volume", the icon with the stroke inside would fade
away when it is disabled.

This commit does the same for "Link ringtone & notification volumes",
making the ringer icon consistent to "Increasing ring volume" when
disabling this option.

This has been tested on Dark Material and has no adverse effects when
making a framework change for Settings. We only need Mute and Audible as
there are only two icon state changes.

Screenshot: http://i.imgur.com/SmJAUWx.png

Thanks to @Spannaa for the idea!

Change-Id: I0c3680a0009c156789b0161e20eb68542a2e8c35

Patchset: 5
http://review.cyanogenmod.org/#/c/102511/