Themes: Don't uninstall theme when processing an overlay fails
Themes can contain overlays for several packages so we should not
uninstall a theme when an overlay fails to be processed.  Simply
log the failure, remove the target from the themes overlay list and
move on to the next.

Change-Id: If701ddbf6e8fc648fa2d715fe8e25cdc5859e05d
(cherry picked from commit 1a665a30bfe172f21ea85d37e18bdd658fed1f77)