Only window position changed, update window crop.
When the window move to the edge of screen, first call setWindowCrop,
if the position of window is changed, second call setPosition.
setWindowCrop will let surfaceflinger dispaly the window, but the
lefttop position is the last postion, this will lead the rightbottom
position reduced, and the reduced rect is dispaly the second layer's
window, so the window flicking.
setPosition let surfaceflinger dispaly the window second, and this
time is right rect.

Solution:  When the window move to the edge of screen, only the
position of window is changed, call setPosition and setWindowCrop.

Change-Id: I235253a02cf7437cfeb9971231ee559d3b8706b3